Chart siirretty componentteihin
Helpottaa charts1 sivun tekoa ja kustomoitua näkymää?
